contacts_api/pyproject.toml
[project]
name = "contacts-api"
version = "0.1.0"
description = ""
authors = [
    {name = "JuliMik",email = "gusachenkojulia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (<4.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "redis (>=5.2.1,<6.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "setuptools (>=79.0.0,<80.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "contacts_api" }]


[tool.poetry.dependencies]
bcrypt = "3.2.0"
